+++ app/templates/gulpfile.js
@@ -88,7 +88,6 @@ gulp.task('concat:js:main', function() {
 			'assets/js/plugins.js',
 			'bower_components/loadcss/src/loadCSS.js',
 			'bower_components/loadcss/src/cssrelpreload.js',
-			'bower_components/fastclick/lib/fastclick.js',
 			'assets/js/main.js'
 		])
 		.pipe($.concat('main.js'))